==> 05_get_match_history.ipynb <==
# Step 5: Collect a player's match history using their PUUID

import pyspark.sql.types as T
import pyspark.sql.functions as F

from pyspark import SparkContext, SparkConf, SQLContext

from dateutil.relativedelta import relativedelta
from pyspark.sql.functions import pandas_udf
from pyspark.sql.window import Window

import pandas as pd
import numpy as np

from datetime import date, datetime, timedelta, timezone
import os
import json
import requests
import time
import yaml
from itertools import chain

appName = "PySpark TFT puuids"
master = "local[*]"
conf = SparkConf() \
    .setAppName(appName) \
    .setMaster(master) \
    .set("spark.executor.memory", "40g") \
    .set("spark.driver.memory", "40g") \
    .set("spark.executor.memoryOverhead", "8g") \
    .set("spark.local.dir", "/home/mai/spark-temp") \
    .set("spark.sql.session.timeZone", "UTC") \
    .set("spark.dynamicAllocation.enabled", "true") \
    .set("spark.dynamicAllocation.minExecutors", "2") \
    .set("spark.dynamicAllocation.maxExecutors", "50") \
    .set("spark.speculation", "true") 
   
sc = SparkContext.getOrCreate(conf=conf)
sqlContext = SQLContext(sc)
spark = sqlContext.sparkSession

## Set up

# retrieve API key
with open('./api_key.yaml', 'r') as file:
    API_KEY = yaml.safe_load(file)['api_key']

## Get match history (most recent 100 matches)
We take the most recent 100 matches and then later we will filter out only the ranked TFT matches

# parameters for API query
REGION = 'kr'
REGION_TO_EXECUTE = 'ASIA'
BASE_URL = f'https://{REGION_TO_EXECUTE}.api.riotgames.com/tft/match/v1/matches/by-puuid'
TIER = 'GOLD'
DIVISION = 'I'
# number of matches per player to retrieve
NUMBER_OF_MATCHES = 100
# set up time limit. We are only interested in matches occured during patch 13.4
PATCH_START_TIME = '1737590400'
PATCH_END_TIME = '1738713599'

# get the puuid of the player
# df = pd.read_parquet(f'./data/tft_puuids/puuids_{REGION}_{TIER}_{DIVISION}.parquet')     # use this path for challenger and grandmaster players
df = pd.read_parquet(f'./data/tft_players/players_{REGION}_{TIER}_{DIVISION}.parquet')     # use this path for gold players

# function to get player's match history based on their puuid
def get_match_history(puuids_, n_matches=100, start=0, end_time=PATCH_END_TIME, start_time=PATCH_START_TIME):
    match_list = []
    i = 0
    while i < len(puuids_):
        # construct url
        url = f'{BASE_URL}/{puuids_[i]}/ids?start={start}&endTime={end_time}&startTime={start_time}&count={n_matches}'
        # headers
        headers = {
            'X-Riot-Token': API_KEY,
        }
        # send request
        response = requests.get(url, headers=headers)
        if response.status_code == 200:
            # parse json response
            print(i)
            data = response.json()
            match_list = list(dict.fromkeys(match_list + data))
            # match_history_data.append(data)
            time.sleep(1.2)
            i += 1
        elif response.status_code == 429:
            print("Rate limit exceeded. Waiting before retrying...")
            # Extract "Retry-After" header if available
            retry_after = int(response.headers.get("Retry-After", 10))
            time.sleep(retry_after)  # Wait for the specified time
        else:
            print(f'Error: {response.status_code} - {response.text}')
            # i += 1
            break
    return match_list

match_hist = get_match_history(puuids)

len(match_hist)

# convert response into a dataframe
match_hist_df = pd.DataFrame(match_hist, columns=['match_id'])
match_hist_df = spark.createDataFrame(match_hist_df)
match_hist_df.show(10)

# write to parquet file
match_hist_df.write.mode('overwrite').parquet(f'./data/tft_matches_id/match_id_{REGION}_{TIER}_{DIVISION}.parquet')
