==> 07_filter_ranked_match_data.ipynb <==
# Step 7: Filter matches
We are only keeping matches where:
* Queue ID = 1100 (ranked queue)
* All 8 players reach carousel round of stage 4 (round 22). This helps eliminate games where there are players who got connection issue and cannot start the game as these players are typically out of the game before that
* All players have at least 1 unit on their board. This eliminates games where players sold all their board and forfeit the game early
* No players are bots. Bot players have their puuid = BOT

import pandas as pd
import numpy as np
import os
import time

import json

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## 1. Load data

# parameters to read the set of match data
REGION = 'kr'
TIER = 'GOLD'
DIVISION = 'I'

dfs = []
for filename in os.listdir('./data/tft_match_data_pandas/'):
    if f'{REGION}_{TIER}_{DIVISION}' in filename:
        df_ = pd.read_parquet(f'./data/tft_match_data_pandas/{filename}')
        dfs.append(df_)

df = pd.concat(dfs, axis=0)

df.head()

# filter for rank matches (queue id = 1100)
df = df.loc[df['info.queueId'] == 1100]

# For each match, we will separate into 8 entries for each player
df = pd.concat([df.reset_index(), pd.json_normalize(df['info.participants'])], axis=1)

df.head()

## 2. Select columns to keep

# select columns to keep
COLS_TO_KEEP = ['metadata.data_version', 'metadata.match_id', 'info.endOfGameResult',
               'info.gameCreation', 'info.game_datetime', 'info.game_length', 'info.game_version',
               'info.mapId', 'info.queueId', 'info.tft_game_type', 'info.tft_set_core_name',
               'info.tft_set_number', 'gold_left', 'last_round', 'level', 'partner_group_id', 'placement',
               'players_eliminated', 'puuid', 'riotIdGameName', 'riotIdTagline',
               'skill_tree', 'time_eliminated', 'total_damage_to_players', 'traits',
               'units', 'win',]

df = df[COLS_TO_KEEP]

df.columns = df.columns.str.replace('.', '_', regex=False)

df.columns

df.head()

## 3. Filter last round of 8th players

# only keep matches where all players reached carousel round of stage 4 (avoid early FF matches)
LAST_ROUND = 22        # the last round of the player who went 8th place has to be >= 22

df['match_earliest_last_round'] = df.groupby('metadata_match_id')['last_round'].transform('min')

df.head()

df = df.loc[df['match_earliest_last_round'] >= 22]

df.shape

df.shape[0] / 8

df.drop('match_earliest_last_round', axis=1, inplace=True)

## 4. Exclude games where some players sell all their board before FF

# some games players will sell all their board before early surrendering, we will also exclude those matches
df['unit_count'] = df['units'].apply(len)
df['match_min_unit_count'] = df.groupby('metadata_match_id')['unit_count'].transform(min)

df.head()

# drop matches where there is a player selling all their units (unit count=0)
df = df.loc[df['match_min_unit_count'] > 0]

df.shape

df.shape[0] / 8

df.drop(['match_min_unit_count', 'unit_count'], axis=1, inplace=True)

## 5. Exclude games with BOTs

df['is_bot'] = (df['puuid'] == 'BOT').astype(int)

df['match_contains_bot'] = df.groupby('metadata_match_id')['is_bot'].transform('max')

# exclude games with bots
df = df.loc[df['match_contains_bot'] == 0]

df.shape

df.shape[0] / 8

df.drop(['is_bot', 'match_contains_bot'], axis=1, inplace=True)

# write to parquet
df.to_parquet(f'./data/tft_match_data_cleaned/matches_{REGION}_{TIER}_{DIVISION}.parquet')
