==> 08_get_reference_data.ipynb <==
# Step 8: Get reference data for set 13
To enrich our dataset, we are using the reference data from Riot's data dragon, downloaded from: https://developer.riotgames.com/docs/tft#data-dragon

This data dragon contains all the reference data for TFT from previous sets as well, so we will have to do some extra work to get only the relevant data for our set (Set 13)

import pandas as pd
import numpy as np

from datetime import date, datetime, timedelta, timezone
import os
import json
import requests
import time
import yaml
from itertools import chain

## 0. Load reference data

DATA_DRAGON_PATH = './data/data_dragon/'
os.makedirs(os.path.join(DATA_DRAGON_PATH, 'set13'), exist_ok=True)

### Units

# champion data
with open(os.path.join(DATA_DRAGON_PATH, 'tft-champion.json'), 'r') as file:
    CHAMP_DICT = json.load(file)

# filter for set 13 units
SET13_CHAMPS = {k.split('/')[-1]: v for (k, v) in CHAMP_DICT['data'].items() if k.startswith('Maps/Shipping/Map22/Sets/TFTSet13/Shop/') or k.startswith('Maps/Shipping/Map22/Sets/TFTSet13_Evolved/Shop/')}

SET13_CHAMPS

with open(os.path.join(DATA_DRAGON_PATH, 'set13', 'units.json'), 'w', encoding='utf-8') as file:
    json.dump(SET13_CHAMPS, file)

### Items
In set 13 we have standard items which include:
* Component items
* Combined items (crafted by combining 2 component items)
* Tactician crown items (crafted by combining 2 spatulas)
* Emblem items (crafted by combining 1 spatula and 1 component item)

We also have these special categories of items are not craftable. They are more powerful than the standard combined items and can only be obtained through loots or augments in the game:
* Artifact items
* Ornn items
* Shimmerscale items
* Radiant items
* Utility items

We also have items specially crafted for Chem Barons units. These items are only obtainable through cashing out Chem Barons and only Chem Baron units can use them

# items data
with open(os.path.join(DATA_DRAGON_PATH, 'tft-item.json'), 'r') as file:
    ITEMS_DICT = json.load(file)

list([i for i in ITEMS_DICT['data'].keys() if '_Ornn' in i])

# filter for set 13 items
def is_set13_items(val):
    if val.startswith('TFT_Item_') and not val.startswith('TFT_Item_Debug'):
        return True
    elif val.startswith('TFT_Consumable_'):
        return True
    elif val.startswith('TFT13_'):
        return True
    elif 'Shimmerscale' in val:
        return True
    elif '_Ornn' in val:
        return True
    elif 'RadiantItems' in val:
        return True
    return False
    
SET13_ITEMS = {k: v for (k, v) in ITEMS_DICT['data'].items() if is_set13_items(k)}

sorted(list(SET13_ITEMS.keys()))

SET13_ITEMS['TFT13_GoopBuff_Miniaturize_Item']

with open(os.path.join(DATA_DRAGON_PATH, 'set13', 'items.json'), 'w', encoding='utf-8') as file:
    json.dump(SET13_ITEMS, file)

### Traits

# traits data
with open(os.path.join(DATA_DRAGON_PATH, 'tft-trait.json'), 'r') as file:
    TRAITS_DICT = json.load(file)

# filter for set 13 traits   
SET13_TRAITS = {k: v for (k, v) in TRAITS_DICT['data'].items() if k.startswith('TFT13_')}

list(SET13_TRAITS.keys())

with open(os.path.join(DATA_DRAGON_PATH, 'set13', 'traits.json'), 'w', encoding='utf-8') as file:
    json.dump(SET13_TRAITS, file)

### Anomalies
Anomaly is a new game mechanic introduced in set 13 that allows players to upgrade one unit on their board with a special effect, making them more powerful.

# items data
with open(os.path.join(DATA_DRAGON_PATH, 'tft-anomalies.json'), 'r') as file:
    ANOMALIES_DICT = json.load(file)

# filter for set 13 anomalies
   
SET13_ANOMALIES = {k: v for (k, v) in ANOMALIES_DICT['data'].items() if k.startswith('TFT13_GoopBuff_')}

list(SET13_ANOMALIES.keys())

with open(os.path.join(DATA_DRAGON_PATH, 'set13', 'anomalies.json'), 'w', encoding='utf-8') as file:
    json.dump(SET13_ANOMALIES, file)
