==> 10_create_match_aggregated_features_and_finalize_dataset.ipynb <==
# Step 10: Create match aggregated features and finalize the dataset
In this notebook, we continue to create match-aggregated features and finalize the dataset:
* Match / performance stats: Create features comparing the player with other opponents (e.g. how many players in the same match reach a higher level or have more gold left at the end)
* Traits, units & items: Create aggregated features of the max/min/avg number of traits/units/items in the match. Then create features comparing the player's count of traits/units/items to those of the opponents.

The final dataset should be structured as follows:
* Index: match id + player id
* Individual performance stats
* Performance stats aggregated at match level
* Number of players with higher performance stats
* List of units with characteristics, number of items equipped, and contesting pressure
* Unit and item count aggregated at player level
* Unit and item count aggregated at match level
* Number of players with equal or higher unit & item count in the same match
* List of active traits with characteristics and contesting pressure
* Active trait count aggregated at player level
* Active trait count aggregated at match level
* Number of players with equal or higher trait count in the same match
* Player's placement in a match and whether or not they were in top 4

import pandas as pd
import numpy as np

import os
import json
import time
import re

import pyspark.sql.types as T
import pyspark.sql.functions as F

from pyspark import SparkContext, SparkConf, SQLContext

from dateutil.relativedelta import relativedelta
from pyspark.sql.functions import pandas_udf
from pyspark.sql.window import Window

from datetime import date, datetime, timedelta, timezone

appName = "PySpark TFT puuids"
master = "local[10]"
conf = SparkConf() \
    .setAppName(appName) \
    .setMaster(master) \
    .set("spark.executor.memory", "40g") \
    .set("spark.driver.memory", "40g") \
    .set("spark.executor.memoryOverhead", "8g") \
    .set("spark.local.dir", "/home/mai/spark-temp") \
    .set("spark.sql.session.timeZone", "UTC") \
    .set("spark.dynamicAllocation.enabled", "true") \
    .set("spark.dynamicAllocation.minExecutors", "2") \
    .set("spark.dynamicAllocation.maxExecutors", "50") \
    .set("spark.speculation", "true") 
   
sc = SparkContext.getOrCreate(conf=conf)
sqlContext = SQLContext(sc)
spark = sqlContext.sparkSession

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

MATCH_DIR = './data/tft_match_data_cleaned/'
MATCH_STAT_DIR = './data/tft_match_stats/'
TRAIT_FEATURE_DIR = './data/tft_trait_features/'
UNIT_ITEM_FEATURE_DIR = './data/tft_unit_item_features/'

# create folders for storing final features
TRAIT_OUTPUT_DIR = './data/all_features/traits/'
os.makedirs(TRAIT_OUTPUT_DIR, exist_ok=True)

UNIT_ITEM_OUTPUT_DIR = './data/all_features/units_items/'
os.makedirs(UNIT_ITEM_OUTPUT_DIR, exist_ok=True)

MATCH_STAT_OUTPUT_DIR = './data/all_features/match_stats'
os.makedirs(MATCH_STAT_OUTPUT_DIR, exist_ok=True)

ALL_FEATURES_DIR = './data/all_features/'

# parameters of region, tier, division
REGION = 'euw1'
TIER = 'GOLD'
DIVISION = 'I'

## 0. Utility functions

### 0.0. Match-specific stats

# function to calculate some extra features on performance statistics in the context of the match
# number of players who reached higher level
# number of players who had more gold left at the end
def create_match_stat_compare_features(df):
    df1 = df.alias('df1')
    df2 = df.alias('df2')

    compare_stat_feats = df1.join(df2, on="metadata_match_id")\
                            .where(F.col('df1.puuid') != F.col('df2.puuid'))\
                            .groupby('df1.metadata_match_id', 'df1.puuid')\
                            .agg((F.sum(F.when(F.col("df2.level") > F.col("df1.level"), 1).otherwise(0))).alias("players_with_higher_level"),
                                 (F.sum(F.when(F.col("df2.gold_left") > F.col("df1.gold_left"), 1).otherwise(0))).alias("players_with_more_gold_left"),
                                )

    return compare_stat_feats

### 0.1. Match-specific trait features

# calculate the max/min/avg for trait count features for each match
def create_match_trait_count_features(df):
    # set up a window for aggregating
    w = Window.partitionBy('metadata_match_id')

    df = df.withColumn('match_avg_active_traits', F.avg('total_active_traits').over(w))\
            .withColumn('match_max_active_traits', F.max('total_active_traits').over(w))\
            .withColumn('match_min_active_traits', F.min('total_active_traits').over(w))\
            .withColumn('match_avg_unique_traits', F.avg('total_unique_traits').over(w))\
            .withColumn('match_max_unique_traits', F.max('total_unique_traits').over(w))\
            .withColumn('match_min_unique_traits', F.min('total_unique_traits').over(w))\
            .withColumn('match_players_with_unique_traits', F.sum(F.when(F.col('total_unique_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_teamup_traits', F.avg('total_teamup_traits').over(w))\
            .withColumn('match_max_teamup_traits', F.max('total_teamup_traits').over(w))\
            .withColumn('match_min_teamup_traits', F.min('total_teamup_traits').over(w))\
            .withColumn('match_players_with_teamup_traits', F.sum(F.when(F.col('total_teamup_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_vertical_traits', F.avg('total_vertical_traits').over(w))\
            .withColumn('match_max_vertical_traits', F.max('total_vertical_traits').over(w))\
            .withColumn('match_min_vertical_traits', F.min('total_vertical_traits').over(w))\
            .withColumn('match_avg_horizontal_traits', F.avg('total_horizontal_traits').over(w))\
            .withColumn('match_max_horizontal_traits', F.max('total_horizontal_traits').over(w))\
            .withColumn('match_min_horizontal_traits', F.min('total_horizontal_traits').over(w))\
            .withColumn('match_avg_defensive_traits', F.avg('total_defensive_traits').over(w))\
            .withColumn('match_max_defensive_traits', F.max('total_defensive_traits').over(w))\
            .withColumn('match_min_defensive_traits', F.min('total_defensive_traits').over(w))\
            .withColumn('match_players_with_defensive_traits', F.sum(F.when(F.col('total_defensive_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_offensive_traits', F.avg('total_offensive_traits').over(w))\
            .withColumn('match_max_offensive_traits', F.max('total_offensive_traits').over(w))\
            .withColumn('match_min_offensive_traits', F.min('total_offensive_traits').over(w))\
            .withColumn('match_players_with_offensive_traits', F.sum(F.when(F.col('total_offensive_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_utility_traits', F.avg('total_utility_traits').over(w))\
            .withColumn('match_max_utility_traits', F.max('total_utility_traits').over(w))\
            .withColumn('match_min_utility_traits', F.min('total_utility_traits').over(w))\
            .withColumn('match_avg_econ_traits', F.avg('total_econ_traits').over(w))\
            .withColumn('match_max_econ_traits', F.max('total_econ_traits').over(w))\
            .withColumn('match_min_econ_traits', F.min('total_econ_traits').over(w))\
            .withColumn('match_players_with_econ_traits', F.sum(F.when(F.col('total_econ_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_bronze_style_traits', F.avg('total_bronze_style_traits').over(w))\
            .withColumn('match_max_bronze_style_traits', F.max('total_bronze_style_traits').over(w))\
            .withColumn('match_min_bronze_style_traits', F.min('total_bronze_style_traits').over(w))\
            .withColumn('match_avg_silver_style_traits', F.avg('total_silver_style_traits').over(w))\
            .withColumn('match_max_silver_style_traits', F.max('total_silver_style_traits').over(w))\
            .withColumn('match_min_silver_style_traits', F.min('total_silver_style_traits').over(w))\
            .withColumn('match_avg_gold_style_traits', F.avg('total_gold_style_traits').over(w))\
            .withColumn('match_max_gold_style_traits', F.max('total_gold_style_traits').over(w))\
            .withColumn('match_min_gold_style_traits', F.min('total_gold_style_traits').over(w))\
            .withColumn('match_avg_prismatic_style_traits', F.avg('total_prismatic_style_traits').over(w))\
            .withColumn('match_max_prismatic_style_traits', F.max('total_prismatic_style_traits').over(w))\
            .withColumn('match_min_prismatic_style_traits', F.min('total_prismatic_style_traits').over(w))\
            .withColumn('match_avg_prismatic_vertical_traits', F.avg('total_prismatic_vertical_traits').over(w))\
            .withColumn('match_max_prismatic_vertical_traits', F.max('total_prismatic_vertical_traits').over(w))\
            .withColumn('match_min_prismatic_vertical_traits', F.min('total_prismatic_vertical_traits').over(w))\
            .withColumn('match_players_with_prismatic_vertical_traits', F.sum(F.when(F.col('total_prismatic_vertical_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_gold_vertical_traits', F.avg('total_gold_vertical_traits').over(w))\
            .withColumn('match_max_gold_vertical_traits', F.max('total_gold_vertical_traits').over(w))\
            .withColumn('match_min_gold_vertical_traits', F.min('total_gold_vertical_traits').over(w))\
            .withColumn('match_players_with_gold_vertical_traits', F.sum(F.when(F.col('total_gold_vertical_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_silver_vertical_traits', F.avg('total_silver_vertical_traits').over(w))\
            .withColumn('match_max_silver_vertical_traits', F.max('total_silver_vertical_traits').over(w))\
            .withColumn('match_min_silver_vertical_traits', F.min('total_silver_vertical_traits').over(w))\
            .withColumn('match_avg_bronze_vertical_traits', F.avg('total_bronze_vertical_traits').over(w))\
            .withColumn('match_max_bronze_vertical_traits', F.max('total_bronze_vertical_traits').over(w))\
            .withColumn('match_min_bronze_vertical_traits', F.min('total_bronze_vertical_traits').over(w))\
            .withColumn('match_avg_prismatic_horizontal_traits', F.avg('total_prismatic_horizontal_traits').over(w))\
            .withColumn('match_max_prismatic_horizontal_traits', F.max('total_prismatic_horizontal_traits').over(w))\
            .withColumn('match_min_prismatic_horizontal_traits', F.min('total_prismatic_horizontal_traits').over(w))\
            .withColumn('match_players_with_prismatic_horizontal_traits', F.sum(F.when(F.col('total_prismatic_horizontal_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_gold_horizontal_traits', F.avg('total_gold_horizontal_traits').over(w))\
            .withColumn('match_max_gold_horizontal_traits', F.max('total_gold_horizontal_traits').over(w))\
            .withColumn('match_min_gold_horizontal_traits', F.min('total_gold_horizontal_traits').over(w))\
            .withColumn('match_players_with_gold_horizontal_traits', F.sum(F.when(F.col('total_gold_horizontal_traits') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_silver_horizontal_traits', F.avg('total_silver_horizontal_traits').over(w))\
            .withColumn('match_max_silver_horizontal_traits', F.max('total_silver_horizontal_traits').over(w))\
            .withColumn('match_min_silver_horizontal_traits', F.min('total_silver_horizontal_traits').over(w))\
            .withColumn('match_avg_bronze_horizontal_traits', F.avg('total_bronze_horizontal_traits').over(w))\
            .withColumn('match_max_bronze_horizontal_traits', F.max('total_bronze_horizontal_traits').over(w))\
            .withColumn('match_min_bronze_horizontal_traits', F.min('total_bronze_horizontal_traits').over(w))
            

    return df

# create features comparing a player's trait count to opponents
def create_match_trait_compare_features(df):
    df1 = df.alias('df1')
    df2 = df.alias('df2')

    compare_trait_feats = df1.join(df2, on="metadata_match_id")\
                            .groupby('df1.metadata_match_id', 'df1.puuid')\
                            .agg((F.sum(F.when(F.col("df2.total_active_traits") > F.col("df1.total_active_traits"), 1).otherwise(0))).alias("players_with_more_active_traits"),
                                (F.sum(F.when(F.col("df2.total_unique_traits") > F.col("df1.total_unique_traits"), 1).otherwise(0))).alias("players_with_more_unique_traits"),
                                (F.sum(F.when(F.col("df2.total_vertical_traits") > F.col("df1.total_vertical_traits"), 1).otherwise(0))).alias("players_with_more_vertical_traits"),
                                (F.sum(F.when(F.col("df2.total_horizontal_traits") > F.col("df1.total_horizontal_traits"), 1).otherwise(0))).alias("players_with_more_horizontal_traits"),
                                (F.sum(F.when(F.col("df2.total_offensive_traits") > F.col("df1.total_offensive_traits"), 1).otherwise(0))).alias("players_with_more_offensive_traits"),
                                (F.sum(F.when(F.col("df2.total_defensive_traits") > F.col("df1.total_defensive_traits"), 1).otherwise(0))).alias("players_with_more_defensive_traits"),
                                (F.sum(F.when(F.col("df2.total_utility_traits") > F.col("df1.total_utility_traits"), 1).otherwise(0))).alias("players_with_more_utility_traits"),
                                (F.sum(F.when(F.col("df2.total_prismatic_style_traits") > F.col("df1.total_prismatic_style_traits"), 1).otherwise(0))).alias("players_with_more_prismatic_style_traits"),
                                (F.sum(F.when(F.col("df2.total_gold_style_traits") > F.col("df1.total_gold_style_traits"), 1).otherwise(0))).alias("players_with_more_gold_style_traits"),
                                (F.sum(F.when(F.col("df2.total_prismatic_vertical_traits") > F.col("df1.total_prismatic_vertical_traits"), 1).otherwise(0))).alias("players_with_more_prismatic_vertical_traits"),
                                (F.sum(F.when(F.col("df2.total_gold_vertical_traits") > F.col("df1.total_gold_vertical_traits"), 1).otherwise(0))).alias("players_with_more_gold_vertical_traits"),
                                (F.sum(F.when(F.col("df2.total_prismatic_horizontal_traits") > F.col("df1.total_prismatic_horizontal_traits"), 1).otherwise(0))).alias("players_with_more_prismatic_horizontal_traits"),
                                (F.sum(F.when(F.col("df2.total_gold_horizontal_traits") > F.col("df1.total_gold_horizontal_traits"), 1).otherwise(0))).alias("players_with_more_gold_horizontal_traits"),)

    return compare_trait_feats

### 0.2. Match-specific unit & item features

# calculate the max/min/avg unit and item count for each match
def create_unit_item_count_features(df):
    w = Window.partitionBy('metadata_match_id')

    df = df.withColumn('match_avg_total_units', F.avg('total_units').over(w))\
            .withColumn('match_max_total_units', F.max('total_units').over(w))\
            .withColumn('match_min_total_units', F.min('total_units').over(w))\
            .withColumn('match_avg_total_items', F.avg('total_items').over(w))\
            .withColumn('match_max_total_items', F.max('total_items').over(w))\
            .withColumn('match_min_total_items', F.min('total_items').over(w))\
            .withColumn('match_avg_total_board_cost', F.avg('total_board_cost').over(w))\
            .withColumn('match_max_total_board_cost', F.max('total_board_cost').over(w))\
            .withColumn('match_min_total_board_cost', F.min('total_board_cost').over(w))\
            .withColumn('match_avg_highest_unit_cost', F.avg('highest_unit_cost').over(w))\
            .withColumn('match_max_highest_unit_cost', F.max('highest_unit_cost').over(w))\
            .withColumn('match_min_highest_unit_cost', F.min('highest_unit_cost').over(w))\
            .withColumn('match_avg_total_ap_units', F.avg('total_ap_units').over(w))\
            .withColumn('match_max_total_ap_units', F.max('total_ap_units').over(w))\
            .withColumn('match_min_total_ap_units', F.min('total_ap_units').over(w))\
            .withColumn('match_avg_total_ad_units', F.avg('total_ad_units').over(w))\
            .withColumn('match_max_total_ad_units', F.max('total_ad_units').over(w))\
            .withColumn('match_min_total_ad_units', F.min('total_ad_units').over(w))\
            .withColumn('match_avg_total_carry_units', F.avg('total_carry_units').over(w))\
            .withColumn('match_max_total_carry_units', F.max('total_carry_units').over(w))\
            .withColumn('match_min_total_carry_units', F.min('total_carry_units').over(w))\
            .withColumn('match_avg_total_caster_units', F.avg('total_caster_units').over(w))\
            .withColumn('match_max_total_caster_units', F.max('total_caster_units').over(w))\
            .withColumn('match_min_total_caster_units', F.min('total_caster_units').over(w))\
            .withColumn('match_avg_total_tank_units', F.avg('total_tank_units').over(w))\
            .withColumn('match_max_total_tank_units', F.max('total_tank_units').over(w))\
            .withColumn('match_min_total_tank_units', F.min('total_tank_units').over(w))\
            .withColumn('match_avg_total_reaper_units', F.avg('total_reaper_units').over(w))\
            .withColumn('match_max_total_reaper_units', F.max('total_reaper_units').over(w))\
            .withColumn('match_min_total_reaper_units', F.min('total_reaper_units').over(w))\
            .withColumn('match_avg_total_fighter_units', F.avg('total_fighter_units').over(w))\
            .withColumn('match_max_total_fighter_units', F.max('total_fighter_units').over(w))\
            .withColumn('match_min_total_fighter_units', F.min('total_fighter_units').over(w))\
            .withColumn('match_avg_total_combined_items', F.avg('total_combined_items').over(w))\
            .withColumn('match_max_total_combined_items', F.max('total_combined_items').over(w))\
            .withColumn('match_min_total_combined_items', F.min('total_combined_items').over(w))\
            .withColumn('match_avg_total_radiant_items', F.avg('total_radiant_items').over(w))\
            .withColumn('match_max_total_radiant_items', F.max('total_radiant_items').over(w))\
            .withColumn('match_min_total_radiant_items', F.min('total_radiant_items').over(w))\
            .withColumn('match_avg_total_tactician_crown_items', F.avg('total_tactician_crown_items').over(w))\
            .withColumn('match_max_total_tactician_crown_items', F.max('total_tactician_crown_items').over(w))\
            .withColumn('match_min_total_tactician_crown_items', F.min('total_tactician_crown_items').over(w))\
            .withColumn('players_with_tactician_crown_items', F.sum(F.when(F.col('total_tactician_crown_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_artifact_items', F.avg('total_artifact_items').over(w))\
            .withColumn('match_max_total_artifact_items', F.max('total_artifact_items').over(w))\
            .withColumn('match_min_total_artifact_items', F.min('total_artifact_items').over(w))\
            .withColumn('players_with_artifact_items', F.sum(F.when(F.col('total_artifact_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_emblem_items', F.avg('total_emblem_items').over(w))\
            .withColumn('match_max_total_emblem_items', F.max('total_emblem_items').over(w))\
            .withColumn('match_min_total_emblem_items', F.min('total_emblem_items').over(w))\
            .withColumn('players_with_emblem_items', F.sum(F.when(F.col('total_emblem_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_ornn_items', F.avg('total_ornn_items').over(w))\
            .withColumn('match_max_total_ornn_items', F.max('total_ornn_items').over(w))\
            .withColumn('match_min_total_ornn_items', F.min('total_ornn_items').over(w))\
            .withColumn('players_with_ornn_items', F.sum(F.when(F.col('total_ornn_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_shimmerscale_items', F.avg('total_shimmerscale_items').over(w))\
            .withColumn('match_max_total_shimmerscale_items', F.max('total_shimmerscale_items').over(w))\
            .withColumn('match_min_total_shimmerscale_items', F.min('total_shimmerscale_items').over(w))\
            .withColumn('players_with_shimmerscale_items', F.sum(F.when(F.col('total_shimmerscale_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_chembaron_items', F.avg('total_chembaron_items').over(w))\
            .withColumn('match_max_total_chembaron_items', F.max('total_chembaron_items').over(w))\
            .withColumn('match_min_total_chembaron_items', F.min('total_chembaron_items').over(w))\
            .withColumn('players_with_chembaron_items', F.sum(F.when(F.col('total_chembaron_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_chembaron_bronze_items', F.avg('total_chembaron_bronze_items').over(w))\
            .withColumn('match_max_total_chembaron_bronze_items', F.max('total_chembaron_bronze_items').over(w))\
            .withColumn('match_min_total_chembaron_bronze_items', F.min('total_chembaron_bronze_items').over(w))\
            .withColumn('players_with_chembaron_bronze_items', F.sum(F.when(F.col('total_chembaron_bronze_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_chembaron_silver_items', F.avg('total_chembaron_silver_items').over(w))\
            .withColumn('match_max_total_chembaron_silver_items', F.max('total_chembaron_silver_items').over(w))\
            .withColumn('match_min_total_chembaron_silver_items', F.min('total_chembaron_silver_items').over(w))\
            .withColumn('players_with_chembaron_silver_items', F.sum(F.when(F.col('total_chembaron_silver_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_chembaron_gold_items', F.avg('total_chembaron_gold_items').over(w))\
            .withColumn('match_max_total_chembaron_gold_items', F.max('total_chembaron_gold_items').over(w))\
            .withColumn('match_min_total_chembaron_gold_items', F.min('total_chembaron_gold_items').over(w))\
            .withColumn('players_with_chembaron_gold_items', F.sum(F.when(F.col('total_chembaron_gold_items') > 0, 1).otherwise(0)).over(w))\
            .withColumn('match_avg_total_chembaron_prismatic_items', F.avg('total_chembaron_prismatic_items').over(w))\
            .withColumn('match_max_total_chembaron_prismatic_items', F.max('total_chembaron_prismatic_items').over(w))\
            .withColumn('match_min_total_chembaron_prismatic_items', F.min('total_chembaron_prismatic_items').over(w))\
            .withColumn('players_with_chembaron_prismatic_items', F.sum(F.when(F.col('total_chembaron_prismatic_items') > 0, 1).otherwise(0)).over(w))

    return df

# create features comparing a player's unit and item count to opponents
def create_match_unit_item_compare_features(df):
    df1 = df.alias('df1')
    df2 = df.alias('df2')

    compare_unit_item_feats = df1.join(df2, on="metadata_match_id")\
                                .groupby('df1.metadata_match_id', 'df1.puuid')\
                                .agg((F.sum(F.when(F.col("df2.total_units") > F.col("df1.total_units"), 1).otherwise(0))).alias("players_with_more_units"),
                                     (F.sum(F.when(F.col("df2.total_board_cost") > F.col("df1.total_board_cost"), 1).otherwise(0))).alias("players_with_higher_total_board_cost"),
                                     (F.sum(F.when(F.col("df2.total_tier_4_units") > F.col("df1.total_tier_4_units"), 1).otherwise(0))).alias("players_with_more_tier_4_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_3_units") > F.col("df1.total_tier_3_units"), 1).otherwise(0))).alias("players_with_more_tier_3_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_2_units") > F.col("df1.total_tier_2_units"), 1).otherwise(0))).alias("players_with_more_tier_2_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_3_1_cost_units") > F.col("df1.total_tier_3_1_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_3_1_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_3_2_cost_units") > F.col("df1.total_tier_3_2_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_3_2_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_3_3_cost_units") > F.col("df1.total_tier_3_3_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_3_3_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_3_4_cost_units") > F.col("df1.total_tier_3_4_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_3_4_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_3_5_cost_units") > F.col("df1.total_tier_3_5_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_3_5_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_3_6_cost_units") > F.col("df1.total_tier_3_6_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_3_6_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_2_1_and_2_cost_units") > F.col("df1.total_tier_2_1_and_2_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_2_1_and_2_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_2_3_cost_units") > F.col("df1.total_tier_2_3_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_2_3_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_2_4_cost_units") > F.col("df1.total_tier_2_4_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_2_4_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_2_5_cost_units") > F.col("df1.total_tier_2_5_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_2_5_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_2_6_cost_units") > F.col("df1.total_tier_2_6_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_2_6_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_1_5_cost_units") > F.col("df1.total_tier_1_5_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_1_5_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_tier_1_6_cost_units") > F.col("df1.total_tier_1_6_cost_units"), 1).otherwise(0))).alias("players_with_more_tier_1_6_cost_units"),
                                     (F.sum(F.when(F.col("df2.total_items") > F.col("df1.total_items"), 1).otherwise(0))).alias("players_with_more_items"),
                                     (F.sum(F.when(F.col("df2.total_combined_items") > F.col("df1.total_combined_items"), 1).otherwise(0))).alias("players_with_more_combined_items"),
                                     (F.sum(F.when(F.col("df2.total_utility_items") > F.col("df1.total_utility_items"), 1).otherwise(0))).alias("players_with_more_utility_items"),
                                     (F.sum(F.when(F.col("df2.total_artifact_items") > F.col("df1.total_artifact_items"), 1).otherwise(0))).alias("players_with_more_artifact_items"),
                                     (F.sum(F.when(F.col("df2.total_ornn_items") > F.col("df1.total_ornn_items"), 1).otherwise(0))).alias("players_with_more_ornn_items"),
                                     (F.sum(F.when(F.col("df2.total_chembaron_items") > F.col("df1.total_chembaron_items"), 1).otherwise(0))).alias("players_with_more_chembaron_items"),
                                     (F.sum(F.when(F.col("df2.total_chembaron_silver_items") > F.col("df1.total_chembaron_silver_items"), 1).otherwise(0))).alias("players_with_more_chembaron_silver_items"),
                                     (F.sum(F.when(F.col("df2.total_chembaron_gold_items") > F.col("df1.total_chembaron_gold_items"), 1).otherwise(0))).alias("players_with_more_chembaron_gold_items"),
                                     (F.sum(F.when(F.col("df2.total_chembaron_prismatic_items") > F.col("df1.total_chembaron_prismatic_items"), 1).otherwise(0))).alias("players_with_more_chembaron_prismatic_items"),
                                    )
    return compare_unit_item_feats

## 1. Load data and create match-aggregated trait features

df = spark.read.parquet(os.path.join(TRAIT_FEATURE_DIR, f'trait_features_{REGION}_{TIER}_{DIVISION}.parquet'))

### 1.1. Match-specific aggregated trait stats

df = create_match_trait_count_features(df)

len(df.columns)

### 1.2. Match-specific trait comparison features

tmp = create_match_trait_compare_features(df)

### 1.3. Join all trait features

df.count()

tmp.count()

df = df.join(tmp, on=['metadata_match_id', 'puuid'], how='left')

df.write.mode('overwrite').parquet(os.path.join(TRAIT_OUTPUT_DIR, f'all_trait_features_{REGION}_{TIER}_{DIVISION}.parquet'))

## 2. Load data and create match-aggregated unit & items features

df = spark.read.parquet(os.path.join(UNIT_ITEM_FEATURE_DIR, f'unit_item_features_{REGION}_{TIER}_{DIVISION}.parquet'))

df.count()

df.printSchema()

### 2.1. Match-specific aggregated unit & item stats

df = create_unit_item_count_features(df)

len(df.columns)

### 2.2. Match-specific unit & item comparison features

tmp = create_match_unit_item_compare_features(df)

# tmp.show(1, False, True)

### 2.3. Join all unit & item features

df.count()

tmp.count()

df = df.join(tmp, on=['metadata_match_id', 'puuid'], how='left')

df.printSchema()

df.write.mode('overwrite').parquet(os.path.join(UNIT_ITEM_OUTPUT_DIR, f'all_unit_item_features_{REGION}_{TIER}_{DIVISION}.parquet'))

## 3. Load data and create match-aggregated performance stats features

df = spark.read.parquet(os.path.join(MATCH_STAT_DIR, f'match_stats_{REGION}_{TIER}_{DIVISION}.parquet')).drop('__index_level_0__')\
            .withColumn('win', F.when(F.col('placement') < 5, 1).otherwise(0))

df.printSchema()

df.count()

### 3.1. Match-specific stat comparison features

tmp = create_match_stat_compare_features(df)

tmp.printSchema()

df.count(), tmp.count()

### 3.2. Join all match features

df = df.join(tmp, on=['metadata_match_id', 'puuid'], how='left')

df.printSchema()

df.write.mode('overwrite').parquet(os.path.join(MATCH_STAT_OUTPUT_DIR, f'all_match_stat_features_{REGION}_{TIER}_{DIVISION}.parquet'))

## 4. Combine all features

df_stat = spark.read.parquet(os.path.join(MATCH_STAT_OUTPUT_DIR, f'all_match_stat_features_{REGION}_{TIER}_{DIVISION}.parquet'))
df_trait = spark.read.parquet(os.path.join(TRAIT_OUTPUT_DIR, f'all_trait_features_{REGION}_{TIER}_{DIVISION}.parquet'))
df_unit = spark.read.parquet(os.path.join(UNIT_ITEM_OUTPUT_DIR, f'all_unit_item_features_{REGION}_{TIER}_{DIVISION}.parquet'))

df_stat.count(), df_trait.count(), df_unit.count()

df_all = df_stat.join(df_trait, on=['metadata_match_id', 'puuid'], how='left')\
                .join(df_unit, on=['metadata_match_id', 'puuid'], how='left')

df_all.count()

df_all.write.mode('overwrite').parquet(os.path.join(ALL_FEATURES_DIR, f'all_features_{REGION}_{TIER}_{DIVISION}.parquet'))

## 5. Check the final schema

df = spark.read.parquet(os.path.join(ALL_FEATURES_DIR, f'all_features_{REGION}_{TIER}_{DIVISION}.parquet'))

df.printSchema()
